Update tests, UG, DG
