+++ docs/DeveloperGuide.adoc
@@ -586,14 +586,49 @@ _Figure 5.5.3.3: Sequence diagram for the execution of `calendar` command_
 
 ==== Command `addEvent`
 Upon entering the `addEvent` command with the correct parameters, an `event` is sent to the `EventsCenter`, which sends
-an update to `CalendarApi` for google calendar to add the event. If the event is added successfully, an event is
+an update to `CalendarApi` for google calendar to add the event. If the calendar event is added successfully, an `event` is
 sent to the `UI` to refresh the calendar. As demonstrated in the diagram below.
 
 image::addEvent.png[width="800"]
 _Figure 5.5.4.1: Sequence diagram for the execution of `addEvent` command_
 
+==== Design Considerations
+**Aspect:** Placement of the `calendarDisplayPanel` +
+**Alternative 1 (current choice):** Replaces the `mainDisplayPanel` +
+**Pros:** Users can see the `PersonListPanel` side by side with the calendar.
+Users can use other functions involving the `PersonListPanel`, such as `add` and `find`, with the calendar open. +
+**Cons:** Calendar is open all the time unless a command changes the `mainDisplayPanel`, more resources consumed as it is
+left running when not needed. +
+**Alternative 2:** An overlay that covers both the `mainDisplayPanel` and the `PersonListPanel` +
+**Pros:** Calendar is switched out when another command is entered, saves resources when not needed. +
+**Cons:** Users are not able to refer to the data on the `PersonListPanel` when using the calendar.
+
 // end::calendar[]
 
+// tag::findfunction[]
+=== Find Mechanism
+==== Description
+The `find` command is the filter function for this application. It will filter through all fields of the user data
+based on the given keywords.
+
+==== Implementation
+The `find` command works as such: `find KEYWORD [MORE KEYWORDS]` +
+
+After the command is entered, the keywords will be sent to the `parser` which then creates a `predicate` to filter
+the stored data. +
+The `predicate` is used by the `model`, which returns a list of filtered people based on the given `predicate`.
+
+==== Design Considerations
+**Aspect:** How to provide the user with functions to search through different fields of customer data +
+**Alternative 1 (current choice):** All-in-one search function. +
+**Pros:** Intuitive and easy to remember. Users can search through any fields of customer data with a single function. +
+**Cons:** User is unable to restrict the search to only one field in the customers' data. E.g. If the user is searching
+for people with addresses in `East Coast`, if there is someone whose name contains `East Coast`, it will show up as well. +
+**Alternative 2:** Multiple search functions. +
+**Pros:** The user can specify exactly which field he wants to search in, avoids the kind of situation described above. +
+**Cons:** Not very intuitive. It is a hassle for the users to remember many different search functions.
+
+// end::findfunction[]
 
 == Documentation
